requirements.txt
django
django-grappelli
pillow
easy_thumbnails
django-modeltranslation
